2023.09.18 [FE] merge solved
병합 후 안맞는 부분 수정
