+++ backend/src/main/java/goormknights/hotel/reservation/model/Reservation.java
@@ -52,7 +52,7 @@ public class Reservation {
     private String notice;              // 고객 요청사항
 
     @ManyToOne
-    @JoinColumn(nullable = false)
+//    @JoinColumn(nullable = false)
     private Item item;
     // 예약 상품 정보: 상품명, 상품 유형, 상품 분류, 기본가, 추가 가능 어른 수, 추가 가능 어린이 수, 어른 추가 비용, 어린이 추가 비용
